Update tea times - from the Master
All of these assumptions that the tea leaves will only be used for a single steeping.

White tea does not need to be steeped for 5 minutes, nor at 175 deg. It is the least processed of all teas, and thus is the most susceptible to improper steeping when water temp is too hot, or for too long.

Boost green tea to 3 min.

Add pu-erh tea times (it's a separate category)

Rooibos is an herbal tea, so remove that.

General time changes otherwise.